Move code that enable tuple comparison in unit tests to new header
